Mobile styles for home page header
