+++ library/src/scss/_coreVariables.scss
@@ -430,6 +430,7 @@ $dropDown-item_minHeight                            : 45px;
 \* -------------------------------------------------------------- */
 
 $vanillaHeader_height                               : 50px !default;
+$vanillaHeader_mobile_height                        : 44px !default;
 $vanillaHeader_fg                                   : #fff !default;
 $vanillaHeader_bg                                   : $global-color_primary !default;
