+++ library/src/scss/components/_meBox.scss
@@ -8,4 +8,8 @@
     display: flex;
     align-items: center;
     height: $vanillaHeader_height;
+
+    @include mediaQuery-panelLayout_oneColumn {
+        height: $vanillaHeader_mobile_height;
+    }
 }
